+++ packages/components/pie-input/src/index.ts
@@ -44,6 +44,21 @@ export class PieInput extends FormControlMixin(RtlMixin(LitElement)) implements 
     @property({ type: Number })
     public maxlength?: InputProps['maxlength'];
 
+    @property({ type: String })
+    public autocomplete?: InputProps['autocomplete'];
+
+    @property({ type: String })
+    public placeholder?: InputProps['placeholder'];
+
+    @property({ type: Boolean })
+    public autoFocus?: InputProps['autoFocus'];
+
+    @property({ type: String })
+    public inputmode?: InputProps['inputmode'];
+
+    @property({ type: Boolean })
+    public readonly?: InputProps['readonly'];
+
     @query('input')
     private input?: HTMLInputElement;
 
@@ -93,7 +108,7 @@ export class PieInput extends FormControlMixin(RtlMixin(LitElement)) implements 
 
     render () {
         const {
-            type, value, name, pattern, minlength, maxlength,
+            type, value, name, pattern, minlength, maxlength, autocomplete, placeholder, autoFocus, inputmode, readonly,
         } = this;
 
         return html`<input
@@ -103,6 +118,11 @@ export class PieInput extends FormControlMixin(RtlMixin(LitElement)) implements 
             pattern=${ifDefined(pattern)}
             minlength=${ifDefined(minlength)}
             maxlength=${ifDefined(maxlength)}
+            autocomplete=${ifDefined(autocomplete)}
+            ?autofocus=${autoFocus}
+            inputmode=${ifDefined(inputmode)}
+            placeholder=${ifDefined(placeholder)}
+            ?readonly=${readonly}
             @input=${this.handleInput}
             @change=${this.handleChange}
             data-test-id="pie-input">`;